+++ test/dummy/db/migrate/20160507024541_add_name_field_to_users.rb
@@ -1,0 +1,5 @@
+class AddNameFieldToUsers < ActiveRecord::Migration
+  def change
+    add_column :users, :name, :string
+  end
+end